Move shallow2 helper into core
