+++ packages/liveblocks-core/src/index.ts
@@ -140,7 +140,7 @@ export { makePoller } from "./lib/Poller";
 export { asPos, makePosition } from "./lib/position";
 export type { Relax } from "./lib/Relax";
 export type { Resolve } from "./lib/Resolve";
-export { shallow } from "./lib/shallow";
+export { shallow, shallow2 } from "./lib/shallow";
 export type { ISignal, SignalType } from "./lib/signals";
 export { batch, DerivedSignal, MutableSignal, Signal } from "./lib/signals";
 export { SortedList } from "./lib/SortedList";
