+++ packages/liveblocks-core/src/lib/__tests__/shallow2.test.ts
@@ -1,6 +1,4 @@
-import { shallow } from "@liveblocks/core";
-
-import { shallow2 } from "../shallow2";
+import { shallow, shallow2 } from "../shallow";
 
 const THING1 = { a: 1 };
 const THING2 = { a: 2 };
